% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clustering_metrics.R
\name{calculate_clustering_metrics}
\alias{calculate_clustering_metrics}
\title{Calculate clustering metrics for a confusion matrix}
\usage{
calculate_clustering_metrics(conf_mat)
}
\arguments{
\item{conf_mat}{a confusion matrix, as produced by \code{\link[c2c]{get_conf_mat}}, or otherwise a confusion matrix of the same form.}
}
\value{
A list containing the metrics that can be calculated, see details.
}
\description{
Calculate a range of clustering metrics on a confusion confusion matrix, usually from \code{\link[c2c]{get_conf_mat}}.
}
\details{
Entropy calculated via \code{\link[c2c]{overall_entropy}} and \code{\link[c2c]{class_entropy}}, purity calculated via \code{\link[c2c]{overall_purity}} and \code{\link[c2c]{class_purity}}, percentage agreement calculated via \code{\link[c2c]{percentage_agreement}} (only for confusion matrcies of equal dimensions)
}
\examples{

## to come
}
\author{
Mitchell Lyons
}
