% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hard.R
\name{get_hard}
\alias{get_hard}
\title{Decompose soft (fuzzy, probabalistic) membership to hard binary matrix}
\usage{
get_hard(x)
}
\arguments{
\item{x}{A matrix or data frame (or something coercible to a matrix) containing memberships - rows are sites (observations, cases etc.) columns are classes}
}
\value{
Binary matrix of class membership. Class names inherited from \code{names()} or \code{colnames()}.
}
\description{
Used in \code{\link[c2c]{get_conf_mat}} but might be useful separately
}
\examples{
my_mat <- matrix(runif(20,0,1), nrow = 4)
get_hard(my_mat)

}
