% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_to_matrix.R
\name{labels_to_matrix}
\alias{labels_to_matrix}
\title{Make a vector of class labels into a hard binary matrix}
\usage{
labels_to_matrix(x)
}
\arguments{
\item{x}{Character or factor vector of class labels}
}
\value{
Binary matrix of class membership.
}
\description{
Used in \code{\link[c2c]{get_conf_mat}} but might be useful separately
}
\examples{
my_labels <- rep(c("a","b","c","d"), 5)
labels_to_matrix(my_labels)

}
