% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_metrics.R
\name{overall_entropy}
\alias{overall_entropy}
\title{Calculate overall cluster entropy}
\usage{
overall_entropy(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix from \code{\link[c2c]{get_conf_mat}} or otherwise (ideally a matrix, although data frames will probably work)}
}
\value{
A scaler, cluster entropy as defined in Manning (2008)
}
\description{
Used to calcualte overall cluster entropy from a confusion matrix.
}
\references{
Manning, C. D., Raghavan, P., & Schütze, H. (2008) Introduction to information retrieval (Vol. 1, No. 1). Cambridge: Cambridge university press.
}
