% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_metrics.R
\name{overall_purity}
\alias{overall_purity}
\title{Calculate overall cluster purity}
\usage{
overall_purity(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix from \code{\link[c2c]{get_conf_mat}} or otherwise (ideally a matrix, although data frames will probably work)}
}
\value{
A scaler, cluster purity as defined in Manning (2008)
}
\description{
Used to calcualte overall cluster purity from a confusion matrix.
}
\references{
Manning, C. D., Raghavan, P., & Schütze, H. (2008) Introduction to information retrieval (Vol. 1, No. 1). Cambridge: Cambridge university press.
}
