% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_metrics.R
\name{percentage_agreement}
\alias{percentage_agreement}
\title{Calculate overall percentage agreement}
\usage{
percentage_agreement(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix from \code{\link[c2c]{get_conf_mat}} or otherwise (ideally a matrix, although data frames will probably work)}
}
\value{
A scaler, percentage agreement (sometime referred to as overall accuracy)
}
\description{
Used to calcualte overall percentage agreement for a confusion matrix - the confusion matrix must have equal dimensions and the diagonal must represent 'matching' class pairs (percentage agreement does not make sense otherwise)
}
